requirements.txt
python-telegram-bot==13.12
urllib3<2.0
standard-imghdr
vk-api==11.9.9
google-cloud-dialogflow==2.25.0
python-dotenv==1.0.0